renamed script for clarity, improved README.md
